+++ Westwind.Utilities/Utilities/StringUtils.cs
@@ -84,10 +84,11 @@ namespace Westwind.Utilities
         /// <param name="value"></param>
         /// <param name="charCount"></param>
         /// <returns></returns>
+        [Obsolete("Please use the StringUtils.Truncate() method instead.")]
         public static string TrimTo(string value, int charCount)
         {
             if (value == null)
-                return string.Empty;
+                return value;
 
             if (value.Length > charCount)
                 return value.Substring(0, charCount);
@@ -118,7 +119,7 @@ namespace Westwind.Utilities
         }
 
         /// <summary>
-        /// Finds the nth index of strting in a string
+        /// Finds the nth index of string in a string
         /// </summary>
         /// <param name="source"></param>
         /// <param name="matchString"></param>
@@ -320,24 +321,6 @@ namespace Westwind.Utilities
             return sb.ToString(); ;
         }
 
-        /// <summary>
-        /// Terminates a string with the given end string/character, but only if the
-        /// value specified doesn't already exist and the string is not empty.
-        /// </summary>
-        /// <param name="value">String to terminate</param>
-        /// <param name="terminator">String to terminate the value string with</param>
-        /// <returns></returns>
-        public static string TerminateString(string value, string terminator)
-        {
-            if (string.IsNullOrEmpty(value))
-                return terminator;
-                    
-            if(value.EndsWith(terminator))
-                return value;
-
-            return value + terminator;
-        }
-
         #endregion
 
         #region String Manipulation
@@ -468,6 +451,18 @@ namespace Westwind.Utilities
         }
 
         /// <summary>
+        /// Truncate a string to maximum length.
+        /// </summary>
+        /// <param name="text">Text to truncate</param>
+        /// <param name="maxLength">Maximum length</param>
+        /// <returns>Trimmed string</returns>
+        public static string Truncate(this string text, int maxLength)
+        {
+            if (string.IsNullOrEmpty(text)) return text;
+            return text.Length <= maxLength ? text : text.Substring(0, maxLength);
+        }
+
+        /// <summary>
         /// Returns an abstract of the provided text by returning up to Length characters
         /// of a text string. If the text is truncated a ... is appended.
         /// </summary>
@@ -486,6 +481,24 @@ namespace Westwind.Utilities
 
             text = text.Substring(0, text.LastIndexOf(" "));
             return text + "...";
+        }
+
+        /// <summary>
+        /// Terminates a string with the given end string/character, but only if the
+        /// text specified doesn't already exist and the string is not empty.
+        /// </summary>
+        /// <param name="value">String to terminate</param>
+        /// <param name="terminator">String to terminate the text string with</param>
+        /// <returns></returns>
+        public static string TerminateString(string value, string terminator)
+        {
+            if (string.IsNullOrEmpty(value))
+                return terminator;
+                    
+            if(value.EndsWith(terminator))
+                return value;
+
+            return value + terminator;
         }
 
         #endregion
@@ -514,6 +527,7 @@ namespace Westwind.Utilities
         {
             return text.IndexOf(searchFor, stringComparison) > -1;
         }
+
 
         /// <summary>
         /// Parses a string into an array of lines broken
@@ -572,7 +586,7 @@ namespace Westwind.Utilities
 
         /// <summary>
         /// Tokenizes a string based on a start and end string. Replaces the values with a token
-        /// value (#@#1#@# for example).
+        /// text (#@#1#@# for example).
         /// 
         /// You can use Detokenize to get the original values back
         /// </summary>
@@ -619,11 +633,11 @@ namespace Westwind.Utilities
         }
 
         /// <summary>
-        /// Parses an string into an integer. If the value can't be parsed
-        /// a default value is returned instead
+        /// Parses an string into an integer. If the text can't be parsed
+        /// a default text is returned instead
         /// </summary>
         /// <param name="input">Input numeric string to be parsed</param>
-        /// <param name="defaultValue">Optional default value if parsing fails</param>
+        /// <param name="defaultValue">Optional default text if parsing fails</param>
         /// <param name="formatProvider">Optional NumberFormat provider. Defaults to current culture's number format</param>
         /// <returns></returns>
         public static int ParseInt(string input, int defaultValue=0, IFormatProvider numberFormat = null)
@@ -640,8 +654,8 @@ namespace Westwind.Utilities
 
 
         /// <summary>
-        /// Parses an string into an decimal. If the value can't be parsed
-        /// a default value is returned instead
+        /// Parses an string into an decimal. If the text can't be parsed
+        /// a default text is returned instead
         /// </summary>
         /// <param name="input"></param>
         /// <param name="defaultValue"></param>
@@ -749,11 +763,11 @@ namespace Westwind.Utilities
         }
 
         /// <summary>
-        /// Retrieves a value by key from a UrlEncoded string.
+        /// Retrieves a text by key from a UrlEncoded string.
         /// </summary>
         /// <param name="urlEncoded">UrlEncoded String</param>
-        /// <param name="key">Key to retrieve value for</param>
-        /// <returns>returns the value or "" if the key is not found or the value is blank</returns>
+        /// <param name="key">Key to retrieve text for</param>
+        /// <returns>returns the text or "" if the key is not found or the text is blank</returns>
         public static string GetUrlEncodedKey(string urlEncoded, string key)
         {
             urlEncoded = "&" + urlEncoded + "&";
@@ -772,10 +786,10 @@ namespace Westwind.Utilities
         }
 
         /// <summary>
-        /// Allows setting of a value in a UrlEncoded string. If the key doesn't exist
-        /// a new one is set, if it exists it's replaced with the new value.
+        /// Allows setting of a text in a UrlEncoded string. If the key doesn't exist
+        /// a new one is set, if it exists it's replaced with the new text.
         /// </summary>
-        /// <param name="urlEncoded">A UrlEncoded string of key value pairs</param>
+        /// <param name="urlEncoded">A UrlEncoded string of key text pairs</param>
         /// <param name="key"></param>
         /// <param name="value"></param>
         /// <returns></returns>
@@ -1051,7 +1065,7 @@ namespace Westwind.Utilities
         }
 
         /// <summary>
-        /// Retrieves a value from an XML-like string
+        /// Retrieves a text from an XML-like string
         /// </summary>
         /// <param name="propertyString"></param>
         /// <param name="key"></param>